Match io.WriterTo signature

+++ html.go
@@ -16,7 +16,7 @@ type HtmlTag struct {
 	*Tag
 }
 
-func (h *HtmlTag) WriteTo(w io.Writer) (int, error) {
+func (h *HtmlTag) WriteTo(w io.Writer) (int64, error) {
 	p, err := nexus.NewPrinter(w)
 	p.Print("<!DOCTYPE html>\n\n")
 	h.Tag.WriteTo(p)
